+++ src/app/admin/dashboard/dashboard.component.ts
@@ -34,6 +34,7 @@ export class DashboardComponent extends HandleSubscription implements OnInit {
       values: [
         {name: 'Users List', icon: 'assets/images/user-gray.svg'},
         {name: 'Publishers', icon: 'assets/images/user-gray.svg'},
+        {name: 'Advertisers', icon: 'assets/images/user-gray.svg'},
         {name: 'Reports', icon: 'assets/images/user-gray.svg'},
       ],
     },