+++ src/app/admin/users/advertiser-list/advertiser-list.component.ts
@@ -1,12 +1,19 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core'
 import { Store } from '@ngrx/store';
 import { animate, style, transition, trigger, } from '@angular/animations';
-import { HandleSubscription } from 'common/handle-subscription';
 import { AppState } from 'models/app-state.model';
-import { Advertisers, AdvertiserInfo } from 'models/settings.model';
-import { sortArrayByKeys } from 'common/utilities/helpers';
-import { TableSortEvent } from 'models/table.model';
 import { LoadAdvertisers } from 'store/admin/admin.actions';
+import { ActivatedRoute, Router } from '@angular/router'
+import {  BaseListComponent } from 'admin/users/base-list/base-list.component'
+
+export interface AdvertisersQueryParams {
+  searchPhrase: string | null,
+  groupBy: 'campaign' | 'user',
+  interval: 'week' | 'day' | 'hour',
+  minDailyViews: number,
+  sort?: string[],
+  order?: 'desc' | 'asc'
+}
 
 @Component({
   selector: 'app-advertiser-list',
@@ -31,81 +38,51 @@ import { LoadAdvertisers } from 'store/admin/admin.actions';
     )
   ],
 })
-export class AdvertiserListComponent extends HandleSubscription implements OnInit {
-  advertisers: Advertisers;
-  filteredAdvertisers: AdvertiserInfo[];
-  searchPhrase = '';
-  groupBy = 'campaign';
-  interval = 'week';
-  minDailyViews = 10000;
-  isLoading: boolean = true;
+export class AdvertiserListComponent extends BaseListComponent implements OnInit {
+  readonly componentStore: Store<AppState>
+  readonly defaultParams: AdvertisersQueryParams = {
+    searchPhrase: '',
+    groupBy: 'campaign',
+    interval: 'week',
+    minDailyViews: 10000,
+  }
+  localStorageName = 'advertisersQueryParams'
 
-  pageSize = 15;
-  page = 1;
-  sortKeys = [];
-  sortDesc = false;
-
-  constructor(private store: Store<AppState>) {
-    super();
+  constructor(
+    store: Store<AppState>,
+    router: Router,
+    activatedRoute: ActivatedRoute
+  ) {
+    super(store, router, activatedRoute);
+    this.componentStore = store
   }
 
-  ngOnInit(): void {
-    const advertisersSubscription = this.store.select('state', 'admin', 'advertisers')
-      .subscribe(advertisers => {
-        this.advertisers = advertisers;
-        this.isLoading = !this.advertisers;
-        this.filterAdvertisers();
-      });
-    this.subscriptions.push(advertisersSubscription);
-    this.loadAdvertisers();
-  }
-
-  loadAdvertisers(): void {
+  loadList(): void {
     this.isLoading = true;
     this.page = 1;
-    this.store.dispatch(new LoadAdvertisers({
-      groupBy: this.groupBy,
-      interval: this.interval,
-      searchPhrase: this.searchPhrase.toLowerCase().trim(),
-      minDailyViews: this.minDailyViews,
+    this.componentStore.dispatch(new LoadAdvertisers({
+      ...this.queryParams
     }));
   }
 
-  groupAdvertisers(groupBy): void {
-    this.groupBy = groupBy;
-    this.loadAdvertisers();
+  get defaultQueryParams (): object {
+    return this.defaultParams
   }
 
-  changeInterval(interval): void {
-    this.interval = interval;
-    this.loadAdvertisers();
-  }
-
-  onSearchChange(): void {
-    this.loadAdvertisers();
-  }
-
-  onMinDailyViewsChange(): void {
-    this.loadAdvertisers();
-  }
-
-  filterAdvertisers(): void {
-    if (this.advertisers && this.advertisers.data) {
-      this.filteredAdvertisers = sortArrayByKeys(this.advertisers.data, this.sortKeys, this.sortDesc);
-      this.filteredAdvertisers = this.filteredAdvertisers.slice((this.page - 1) * this.pageSize, this.page * this.pageSize)
-    } else {
-      this.filteredAdvertisers = [];
+  ngOnInit(): void {
+    const advertisersSubscription = this.componentStore.select('state', 'admin', 'advertisers')
+      .subscribe(advertisers => {
+        this.list = advertisers;
+        this.isLoading = !this.list;
+        this.onPageChange();
+      });
+    this.subscriptions.push(advertisersSubscription);
+    this.queryParams = {
+      ...this.defaultParams,
+      sort: this.sortKeys[0] || null,
+      order: this.sortDesc ? 'desc' : 'asc',
     }
-  }
-
-  sortTable(event: TableSortEvent): void {
-    this.sortKeys = event.keys;
-    this.sortDesc = event.sortDesc;
-    this.filterAdvertisers();
-  }
-
-  handlePaginationEvent(e): void {
-    this.page = e.pageIndex + 1;
-    this.filterAdvertisers();
+    this.subscriptions.push(this.checkQueryParams())
+    this.loadList();
   }
 }