+++ src/app/admin/users/user-list/user-list.component.ts
@@ -1,13 +1,22 @@
 import { Component, OnInit } from '@angular/core'
 import { Store } from '@ngrx/store'
 import { animate, style, transition, trigger } from '@angular/animations'
-import { HandleSubscription } from 'common/handle-subscription'
 import { AppState } from 'models/app-state.model'
-import { Users } from 'models/settings.model'
-import { TableSortEvent } from 'models/table.model'
-import * as adminActions from 'store/admin/admin.actions'
 import { appSettings } from 'app-settings'
 import { SessionService } from '../../../session.service'
+import { ActivatedRoute, Router } from '@angular/router'
+import { BaseListComponent } from 'admin/users/base-list/base-list.component'
+import { LoadUsers } from 'store/admin/admin.actions'
+
+export interface UsersQueryParams {
+  selectedType: 'Advertisers' | 'Publishers' | 'All'
+  onlyEmailUnconfirmed: boolean,
+  onlyAdminUnconfirmed: boolean,
+  userSearch: string | null
+  sort?: string[],
+  order?: 'desc' | 'asc'
+}
+
 
 @Component({
   selector: 'app-user-list',
@@ -32,91 +41,71 @@ import { SessionService } from '../../../session.service'
     ),
   ],
 })
-export class UserListComponent extends HandleSubscription implements OnInit {
-  users: Users
-  isLoading: boolean = true
+export class UserListComponent extends BaseListComponent implements OnInit {
+  readonly componentStore: Store<AppState>
+  readonly defaultParams: UsersQueryParams = {
+    selectedType: 'All',
+    onlyEmailUnconfirmed: false,
+    onlyAdminUnconfirmed: false,
+    userSearch: null
+  }
+  localStorageName = 'usersQueryParams'
   userTypes: string[] = appSettings.USER_TYPES
-  selectedType: string = 'All'
-  onlyEmailUnconfirmed: boolean = false
-  onlyAdminUnconfirmed: boolean = false
-  userSearch: string = null
-  sortKeys: string[] = []
-  sortDesc: boolean = false
 
-  constructor (
-    private store: Store<AppState>,
-    private session: SessionService
+  constructor(
+    store: Store<AppState>,
+    router: Router,
+    activatedRoute: ActivatedRoute,
+    private session: SessionService,
   ) {
-    super()
+    super(store, router, activatedRoute);
+    this.componentStore = store
+  }
+
+  loadList(nextPage?: string): void {
+    this.isLoading = true;
+    const filters = [];
+    this.page = 1;
+    if (this.queryParams.onlyEmailUnconfirmed) {
+      filters.push('email-unconfirmed');
+    }
+    if (this.queryParams.onlyAdminUnconfirmed) {
+      filters.push('admin-unconfirmed');
+    }
+    if (this.queryParams.selectedType) {
+      filters.push(this.queryParams.selectedType.toLowerCase());
+    }
+    this.componentStore.dispatch(new LoadUsers({
+      nextPage,
+      searchPhrase: this.queryParams.userSearch ? this.queryParams.userSearch.toLowerCase().trim() : null,
+      filters,
+      orderBy: this.queryParams.sort,
+      direction: this.queryParams.order,
+    }));
+  }
+
+  get defaultQueryParams (): object {
+    return this.defaultParams
   }
 
   ngOnInit () {
-    const usersSubscription = this.store.select('state', 'admin', 'users').
+    const usersSubscription = this.componentStore.select('state', 'admin', 'users').
       subscribe(users => {
-        this.users = users
-        this.isLoading = !this.users
+        this.list = users
+        this.isLoading = !this.list
       })
     this.subscriptions.push(usersSubscription)
-    this.loadUsers()
-  }
-
-  loadUsers (nextPage ?: string): void {
-    this.isLoading = true
-    const filters = [];
-    if (this.onlyEmailUnconfirmed) {
-      filters.push('email-unconfirmed');
+    this.queryParams = {
+      ...this.defaultParams,
+      sort: this.sortKeys[0] || null,
+      order: this.sortDesc ? 'desc' : 'asc',
     }
-    if (this.onlyAdminUnconfirmed) {
-      filters.push('admin-unconfirmed');
-    }
-    if (this.selectedType) {
-      filters.push(this.selectedType.toLowerCase());
-    }
-    this.store.dispatch(new adminActions.LoadUsers({
-      nextPage,
-      searchPhrase: this.userSearch ? this.userSearch.toLowerCase().trim() : null,
-      filters,
-      orderBy: this.sortKeys.join(','),
-      direction: this.sortDesc ? 'desc' : 'asc',
-    }))
+    this.subscriptions.push(this.checkQueryParams())
+    this.loadList()
+
   }
 
-  filterUsersByType (type, resetSearch = false) {
-    this.selectedType = type
-    if (resetSearch) {
-      this.userSearch = null
-      this.loadUsers()
-    }
-  }
-
-  filterEmailUnconfirmed () {
-    this.onlyEmailUnconfirmed = !this.onlyEmailUnconfirmed;
-    this.loadUsers()
-  }
-
-  filterAdminUnconfirmed () {
-    this.onlyAdminUnconfirmed = !this.onlyAdminUnconfirmed;
-    this.loadUsers()
-  }
-
-  onSearchChange () {
-    this.loadUsers()
-  }
-
-  sortTable (event: TableSortEvent) {
-    this.sortKeys = event.keys
-    this.sortDesc = event.sortDesc
-    this.loadUsers()
-  }
-
-  get showActions () : boolean {
+  get showActions() : boolean {
     return this.session.isModerator()
-  }
-
-  handlePaginationEvent (e): void {
-    const payload = this.users.prevPageUrl && this.users.currentPage >=
-    e.pageIndex + 1 ? this.users.prevPageUrl
-      : this.users.nextPageUrl
-    this.loadUsers(payload)
   }
 }